fix upConv usage
As the help of `upConv` says:
```
>> help upConv
  RES = upConv(IM, FILT, EDGES, STEP, START, STOP, RES)

  ...
  RES is an optional result matrix.  The convolution result will be
  destructively added into this matrix.  If this argument is passed, the
  result matrix will not be returned. DO NOT USE THIS ARGUMENT IF
  YOU DO NOT UNDERSTAND WHAT THIS MEANS!!
  ...
```

So, what you can do here is to modify the code:

1. Not to pass the `RES` input argument
2. Not to receive the `RES` output argument

The problem with the following code is that `res` exists both on the left side and right side.

https://github.com/LabForComputationalVision/matlabPyrTools/blob/42e4602dd06a36282b16123e7f587c08f397b272/reconSpyr.m#L94

According to the `upConv` usage in reconSpyr.m, and reconSpyrLevs.m, the `res` variable is created in proper size and type, so you should omit the left side `res` output argument.